# fixed bug in fixed candidate ranking and add speedup option for ranking eval