**Patch description**
1) fix error setting up fixed candidates
~~2) separate loss instantiation into separate function~~ -> moved to #1899
3) add rank_top_k option to do top-k selection instead of sorting all candidates by score. if you're just looking for hits@ metrics this is much faster, though this makes mean_rank / mrr calculations a bit wonky (it doesn't sort anything beyond the k). I'm open to suggestions on how to handle this--maybe good to zero them out?
4) switch pad handling to be a bit faster when using inline candidates


**Testing steps**
The following run uses a relatively large 75k candidate set with fixed encodings during evaluation. I added a timer around valid/test in train_mode.py.
```
$ python examples/train_model.py -t wikimovies --dict-file /tmp/dict_wikimovies -bs 128 -nt 10 -eps 5 -m memnn --no-cuda -cands batch -ecands fixed -fcp /private/home/ahm/ParlAI/data/WikiMovies/movieqa/knowledge_source/entities.txt -dt train:ordered --ignore-bad-candidates True -mf /tmp/run
```
master: `~48s training / IndexError on eval`
master + error fix: `~48s training / ~49s valid + test`
rank_speed: `~46s training / ~47s valid + test` (slightly faster training + slightly faster eval)
rank_top_k=100: `~46s training / ~36s valid + test` (faster eval if you just want hits@100)
rank_top_k=1: `~47s training / ~32s valid + test` (faster eval if you just want hits@1)

The biggest speedup you'll see is when using inline candidates due to the change in the "ordering" code.
```
python examples/train_model.py -t wikimovies --dict-file /tmp/dict_wikimovies -bs 4 -esz 1 -hops 0 --memsize 0 -tf False -nt 10 -eps 1 -m memnn --no-cuda -cands batch -ecands inline -dt train:ordered -vme 1000 --short-final-eval True -mf /tmp/run
```
master: `467s valid + test`
rank_speed: `402s valid + test`
rank_top_k=100: `388s valid + test`

You might notice that this is a no-hop memnn with esz of only 1, and it still takes forever to do just 1000 examples here. Precomputing the candidate vectors helps a lot. (Or even better, the candidate vectors.)