@alexholdenmiller 2019-08-06
So this passes all existing tests but it would be good to know: what are some other fixed-candidate-ranking scenarios with either different models or different tasks?

I switched to a transformer and tried this:
```
python examples/train_model.py -t wikimovies --dict-file /tmp/dict_wikimovies -bs 256 -nt 1 -eps 1 -cands batch -ecands fixed -fcp /private/home/ahm/ParlAI/data/WikiMovies/movieqa/knowledge_source/entities.txt -dt train:ordered --ignore-bad-candidates True -m transformer/ranker -mf /tmp/run
```
master: `~33s train / IndexError on eval`
rank_speed: `~33s train / no error on eval`